Changes pattern to using interfaces.
Interfaces drastically improve compilation time and IDE responsiveness as it prevents compiling the world when you only want to make small changes.  It also will enable us to create Mocks of most of our contracts which should facilitate unit testing, without the need to construct the world.

The interfaces _do_ result in some redundancies and maintenance burden, but I believe the performance and testing benefits outweigh those costs.

There may be a couple contracts that don't follow the interface pattern yet, and that is fine and we can address when the time comes.

Other changes that came up while doing this:
* extracted out Controlled from Controller
* removed some unused parameters from Orders.getOrders and OrdersFetcher.getOrder
* refactored OrderBook.getOrderBook to address stack depth problems
* deleted Interfaces.sol, it was meant to be temporary and this extracts out the last couple interfaces
* changed isContainerFor* methods in Market to match patterns in Branch
* moved `Trading.sol` to `trading` folder from libraries